+++ src/numpydantic/__init__.py
@@ -5,7 +5,8 @@
 from numpydantic.ndarray import NDArray
 from numpydantic.meta import update_ndarray_stub
 from numpydantic.validation.shape import Shape
+from numpydantic.annotation import NDArraySchema
 
 update_ndarray_stub()
 
-__all__ = ["NDArray", "Shape"]
+__all__ = ["NDArray", "NDArraySchema", "Shape"]